% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{adjmatrix}
\alias{adjmatrix}
\title{Function to extract the adjacency matrix form a cw object}
\usage{
adjmatrix(x)
}
\arguments{
\item{x}{object of class cw}
}
\description{
Function to extract the adjacency matrix form a cw object
}

