% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmc_prep-data.R
\name{cmc_prep}
\alias{cmc_prep}
\title{Variable frame for a three block model for the CMC data}
\format{A 5x3 variable frame}
\description{
Block dependencies are Block 1 <- Block 2 <- Block 3
}
\details{
Block 1 (targets) : contraceptive <-> nrChild
Block 2 (intermediate variables): wifeRel <-> solIndex
Block 3 (purely explanatory): age
}

