% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{coxwer}
\alias{coxwer}
\title{A function to fit a block recursive chain graph model in R following the outlines of Cox & Wermuth 1996 and Caputo et al 1997; some adaptions made by us}
\usage{
coxwer(var.frame, data, autodetect = FALSE, pen, signif = 0.01,
  contrasts = c("contr.treatment", "contr.treatment"),
  scalepredictors = c("no", "center", "scale", "normalize"), silent = FALSE,
  restr = NULL, adjfile = NULL, ...)
}
\arguments{
\item{var.frame}{A variable.frame with the following structure: First column are the variable names, second column are the variable types (one of categorical, ordinal, continuous, binary, count, odcount (overdispersed count), gamma or invgaussian, see also \code{\link{prep_coxwer}}), third column are the block which are labeled from left to right increasingly, so the left most block is 1 and then 2 and so forth; the block with purely dependent variable is block 1 the block with purely explanatory variables is the right most block. For an exmaple see \code{\link{cmc_prep}}}

\item{data}{the data frame conating the varibales in var.frame}

\item{autodetect}{flag whether the function should derive the variable type for the R object (TRUE if so); will override user specifications in var.frame; defaults to FALSE}

\item{pen}{the penalty applied to the information criterion for variable selection; defaults to #parameters*log(n) (BIC); -2 would be AIC}

\item{signif}{the significance level used in the tests for inclusion of higher order effects; defaults to 0.01}

\item{contrasts}{which contrasts to use for ordinal and nominal predictors; defaults to treatment contrasts all around}

\item{scalepredictors}{whtehr and how to scale the predictors (no=no scaling, center=mean centering, scale=standardization (mean=0 and var=1) and normalize=scaling to [0,1])}

\item{silent}{flags whether output should be printed (if TRUE, no output; the default)}

\item{restr}{a matrix of the same size as the adjacency matrix in the model so as to supress specific dependencies fitted in the algorithm (0 means not fitted)}

\item{adjfile}{if TRUE the adjacency matrix is written to a file}

\item{...}{additional arguments (not used currently)}
}
\value{
an object of class cw which contains a list of fitted models, the variable frame and the fitted adjacency matrix
}
\description{
A function to fit a block recursive chain graph model in R following the outlines of Cox & Wermuth 1996 and Caputo et al 1997; some adaptions made by us
}
\seealso{
\code{\link{prep_coxwer}} \code{\link{cmc_prep}}
}

