% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{fitfunc}
\alias{fitfunc}
\title{the fitting wrapper for coxwer}
\usage{
fitfunc(fmla, data = data, type, currfamily = currfamily, ...)
}
\arguments{
\item{fmla}{an R formula}

\item{data}{the data frame to take the variables from}

\item{type}{what is the type of the target variable}

\item{currfamily}{current family for the glm}

\item{...}{further arguents passed to the fitting functions}
}
\value{
a fitted model of either class multinom (categorical type) or polr (ordinal type) or glm.nb (count type) or glm (for type being metric or positive or so)
}
\description{
the fitting wrapper for coxwer
}

