% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{plot.cw}
\alias{plot.cw}
\title{S3 plot method;  Interactively plot the coxwer object}
\usage{
\method{plot}{cw}(x, interactive = TRUE, ...)
}
\arguments{
\item{x}{an object of class cw}

\item{interactive}{should the plot be interactive (defaults to TRUE)}

\item{...}{further arguments to tkplot (if interactive) or to plot.igraph (if not interactive)}
}
\description{
S3 plot method;  Interactively plot the coxwer object
}

