% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{predict.cw}
\alias{predict.cw}
\title{S3 predict method}
\usage{
\method{predict}{cw}(object, target, newdata, type, silent = FALSE, ...)
}
\arguments{
\item{object}{a cw object}

\item{target}{Which target variable to predict? Defaults to NULL which predicts all}

\item{newdata}{new data frame}

\item{type}{what type of predictions are wanted (if missing it is "")}

\item{silent}{Should info be printed (FALSE means yes)}

\item{...}{additional arguents passed to predict for the model classes}
}
\description{
S3 predict method
}

