% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{prep_coxwer}
\alias{prep_coxwer}
\title{Function to interactively prepare a variable frame for coxwer}
\usage{
prep_coxwer(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a variable frame
}
\description{
Function to interactively prepare a variable frame for coxwer
}

