% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{print.cw}
\alias{print.cw}
\title{S3 print method (prints adjacency matrix)}
\usage{
\method{print}{cw}(x, ...)
}
\arguments{
\item{x}{object of class cw}

\item{...}{Further arguments (no effect)}
}
\description{
S3 print method (prints adjacency matrix)
}

