% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{summary.cw}
\alias{summary.cw}
\title{S3 summary method}
\usage{
\method{summary}{cw}(object, target = NULL, vframe = FALSE, adj = FALSE,
  ...)
}
\arguments{
\item{object}{a cw object}

\item{target}{summary for which target variable, "all" gives all models? Defaults to NULL which gives nothing models}

\item{vframe}{flags whether variable frame should be summarized, defaults to FALSE}

\item{adj}{flags whether adjacency matrix should be summarized, defaults to FALSE}

\item{...}{further args}
}
\description{
S3 summary method
}

