% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{write_cw}
\alias{write_cw}
\title{Write the coxwer adjacency matrix plot to a file}
\usage{
write_cw(obj, file, ...)
}
\arguments{
\item{obj}{an object of class cw}

\item{file}{filename}

\item{...}{further arguments to write.graph}
}
\description{
Write the coxwer adjacency matrix plot to a file
}

