% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxwer.R
\name{plot.cw}
\alias{plot.cw}
\title{S3 plot method;  Interactively plot the coxwer object}
\usage{
\method{plot}{cw}(x, ...)
}
\arguments{
\item{x}{an object of class cw}

\item{...}{further arguments to tkplot}
}
\description{
S3 plot method;  Interactively plot the coxwer object
}

